+++ crates/rhdl-fpga/src/serial_bus/mod.rs
@@ -31,6 +31,7 @@ pub mod one_wire_master;
 pub mod rs485_master;
 pub mod sent_rx;
 pub mod smbus_host;
+pub mod smpte_ltc_encoder;
 pub mod spi_master;
 pub mod spi_slave;
 pub mod ti_hdq;